Update distance.Rmd
The edit might look like it's nothing, but I actually replaced the left-quotes with regular quotes on the two image html tags at the beggining of the file, because they were causing a bug which made the HTML tag escaped after the .rmd was knit (tested on IE and Chrome). Note that the bug is not present at the .Rmd preview here on Github. Just check the current book status on http://genomicsclass.github.io/book/pages/distance.html, the first two images aren't shown. Also, the width parameters did not have the quotes. I made some testing and this should fix the problem.